# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Emory",
        "description": "http://127.0.0.1:8000/api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/resources": {
            "get": {
                "tags": [
                    "Resource Management"
                ],
                "summary": "Resource Management",
                "description": " Get all resource",
                "operationId": "Resources",
                "responses": {
                    "200": {
                        "description": "Data fetched",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/resource/{id}/details": {
            "get": {
                "tags": [
                    "Resource Management"
                ],
                "summary": "Get details of particular resource",
                "description": "Get Get details of particular resource",
                "operationId": "Resource",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "resource id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details fetched",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}